+++ timeboost-sequencer/src/decrypt.rs
@@ -375,7 +375,7 @@ impl Worker {
             .lock()
             .iter()
             .filter(|(k, _)| k.round() == round)
-            .all(|(_, v)| usize::from(self.committee.threshold().get()) < v.len());
+            .all(|(_, v)| self.committee.threshold().get() < v.len());
 
         if !hatched {
             // ciphertexts are not ready to be decrypted.
